refactor: modularize codebase into core, plotting, and io modules
